requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
@EXPLICIT
https://repo.anaconda.com/pkgs/main/osx-64/blas-1.0-mkl.conda
https://conda.anaconda.org/conda-forge/osx-64/ca-certificates-2020.12.5-h033912b_0.tar.bz2
https://repo.anaconda.com/pkgs/main/osx-64/intel-openmp-2019.4-233.conda
https://repo.anaconda.com/pkgs/main/osx-64/jpeg-9b-he5867d9_2.conda
https://repo.anaconda.com/pkgs/main/osx-64/libcxx-10.0.0-1.conda
https://repo.anaconda.com/pkgs/main/osx-64/libgfortran-3.0.1-h93005f0_2.conda
https://repo.anaconda.com/pkgs/main/osx-64/libuv-1.40.0-haf1e3a3_0.conda
https://repo.anaconda.com/pkgs/main/osx-64/llvm-openmp-10.0.0-h28b9765_0.conda
https://repo.anaconda.com/pkgs/main/osx-64/xz-5.2.5-h1de35cc_0.conda
https://repo.anaconda.com/pkgs/main/osx-64/zlib-1.2.11-h1de35cc_3.conda
https://repo.anaconda.com/pkgs/main/osx-64/libffi-3.3-hb1e8313_2.conda
https://repo.anaconda.com/pkgs/main/osx-64/libpng-1.6.37-ha441bb4_0.conda
https://repo.anaconda.com/pkgs/main/osx-64/lz4-c-1.9.3-h23ab428_0.conda
https://repo.anaconda.com/pkgs/main/osx-64/mkl-2019.4-233.conda
https://repo.anaconda.com/pkgs/main/osx-64/ncurses-6.2-h0a44026_1.conda
https://conda.anaconda.org/conda-forge/osx-64/openssl-1.1.1i-h35c211d_0.tar.bz2
https://repo.anaconda.com/pkgs/main/osx-64/tk-8.6.10-hb0a8c7a_0.conda
https://repo.anaconda.com/pkgs/main/osx-64/freetype-2.10.4-ha233b18_0.conda
https://repo.anaconda.com/pkgs/main/osx-64/libedit-3.1.20191231-h1de35cc_1.conda
https://repo.anaconda.com/pkgs/main/osx-64/readline-8.1-h9ed2024_0.conda
https://repo.anaconda.com/pkgs/main/osx-64/zstd-1.4.5-h41d2c2f_0.conda
https://repo.anaconda.com/pkgs/main/osx-64/libtiff-4.1.0-hcb84e12_1.conda
https://repo.anaconda.com/pkgs/main/osx-64/sqlite-3.33.0-hffcf06c_0.conda
https://repo.anaconda.com/pkgs/main/osx-64/lcms2-2.11-h92f6f08_0.conda
https://repo.anaconda.com/pkgs/main/osx-64/python-3.8.5-h26836e1_1.conda
https://conda.anaconda.org/anaconda/noarch/colorama-0.4.4-py_0.tar.bz2
https://repo.anaconda.com/pkgs/main/osx-64/kiwisolver-1.3.1-py38h23ab428_0.conda
https://repo.anaconda.com/pkgs/main/osx-64/ninja-1.10.2-py38hf7b0b51_0.conda
https://repo.anaconda.com/pkgs/main/noarch/olefile-0.46-py_0.conda
https://repo.anaconda.com/pkgs/main/noarch/pyparsing-2.4.7-pyhd3eb1b0_0.tar.bz2
https://conda.anaconda.org/conda-forge/osx-64/python_abi-3.8-1_cp38.tar.bz2
https://repo.anaconda.com/pkgs/main/osx-64/six-1.15.0-py38hecd8cb5_0.conda
https://repo.anaconda.com/pkgs/main/osx-64/termcolor-1.1.0-py38_1.conda
https://repo.anaconda.com/pkgs/main/noarch/threadpoolctl-2.1.0-pyh5ca1d4c_0.conda
https://repo.anaconda.com/pkgs/main/osx-64/tornado-6.1-py38h9ed2024_0.conda
https://repo.anaconda.com/pkgs/main/noarch/typing_extensions-3.7.4.3-pyh06a4308_0.conda
https://repo.anaconda.com/pkgs/main/noarch/wheel-0.36.2-pyhd3eb1b0_0.conda
https://repo.anaconda.com/pkgs/main/noarch/xmltodict-0.12.0-py_0.conda
https://conda.anaconda.org/conda-forge/osx-64/certifi-2020.12.5-py38h50d1736_1.tar.bz2
https://repo.anaconda.com/pkgs/main/osx-64/cycler-0.10.0-py38_0.conda
https://repo.anaconda.com/pkgs/main/osx-64/mkl-service-2.3.0-py38h9ed2024_0.conda
https://conda.anaconda.org/conda-forge/osx-64/orjson-3.4.8-py38hca655e8_0.tar.bz2
https://repo.anaconda.com/pkgs/main/osx-64/pillow-8.1.0-py38h5270095_0.conda
https://repo.anaconda.com/pkgs/main/noarch/python-dateutil-2.8.1-pyhd3eb1b0_0.tar.bz2
https://repo.anaconda.com/pkgs/main/osx-64/numpy-base-1.19.2-py38hcfb5961_0.conda
https://repo.anaconda.com/pkgs/main/osx-64/setuptools-52.0.0-py38hecd8cb5_0.conda
https://repo.anaconda.com/pkgs/main/noarch/joblib-1.0.0-pyhd3eb1b0_0.conda
https://repo.anaconda.com/pkgs/main/osx-64/pip-20.3.3-py38hecd8cb5_0.conda
https://repo.anaconda.com/pkgs/main/osx-64/matplotlib-3.3.2-hecd8cb5_0.conda
https://repo.anaconda.com/pkgs/main/osx-64/matplotlib-base-3.3.2-py38h181983e_0.conda
https://repo.anaconda.com/pkgs/main/osx-64/mkl_fft-1.2.0-py38hc64f4ea_0.conda
https://repo.anaconda.com/pkgs/main/osx-64/mkl_random-1.1.1-py38h959d312_0.conda
https://repo.anaconda.com/pkgs/main/osx-64/numpy-1.19.2-py38h456fd55_0.conda
https://conda.anaconda.org/pytorch/osx-64/pytorch-1.7.1-py3.8_0.tar.bz2
https://repo.anaconda.com/pkgs/main/osx-64/scipy-1.5.2-py38h2515648_0.conda
https://repo.anaconda.com/pkgs/main/osx-64/scikit-learn-0.23.2-py38h959d312_0.conda
https://conda.anaconda.org/pytorch/osx-64/torchaudio-0.7.2-py38.tar.bz2
https://conda.anaconda.org/pytorch/osx-64/torchvision-0.8.2-py38_cpu.tar.bz2

numpy~=1.19.2
orjson~=3.4.8
matplotlib~=3.3.2
pillow~=8.1.0
scikit-learn~=0.23.2